Remove kairos-prover-server-lib and describe lsp fix that makes it unnecessary
