+++ Cargo.toml
@@ -9,7 +9,6 @@ members = [
   "kairos-tx",
   "kairos-test-utils",
   "kairos-circuit-logic",
-  "kairos-prover-server-lib",
 ]
 
 [workspace.package]
