+++ flake.nix
@@ -60,7 +60,6 @@
                   "/kairos-test-utils"
                   "/kairos-tx"
                   "/kairos-circuit-logic"
-                  "/kairos-prover-server-lib"
                   "/Cargo.toml"
                   "/Cargo.lock"
                 ]) && (
