--- kairos-prover-server-lib/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "kairos-prover-server-lib"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[features]
-arbitrary = ["kairos-circuit-logic/arbitrary"]
-
-[profile.test]
-opt-level = 3
-debug-assertions = true
-overflow-checks = true
-
-[dependencies]
-dotenvy = "0.15"
-axum = { version = "0.7", features = ["tracing"] }
-axum-extra = { version = "0.9", features = [
-    "typed-routing",
-    "typed-header",
-    "json-deserializer",
-] }
-serde = { version = "1", features = ["derive"] }
-# TODO switch to bincode 2
-serde_json = "1.0"
-tokio = { version = "1", features = ["tracing"] }
-tracing = "0.1"
-kairos-circuit-logic = { path = "../kairos-circuit-logic", features = ["serde"] }
-kairos-trie = { git = "https://github.com/cspr-rad/kairos-trie", features = ["serde"] }
