+++ kairos-prover/kairos-prover-risc0-server/src/main.rs
@@ -1,4 +1,7 @@
-use kairos_prover_server_lib::{run_server, Proof, Prover};
+use axum::{http::StatusCode, Json};
+use axum_extra::routing::{RouterExt, TypedPath};
+use kairos_circuit_logic::{ProofInputs, ProofOutputs};
+
 // These constants represent the RISC-V ELF and the image ID generated by risc0-build.
 // The ELF is used for proving and the ID is used for verification.
 use methods::{PROVE_BATCH_ELF, PROVE_BATCH_ID};
@@ -6,48 +9,76 @@ use risc0_zkvm::{ExecutorEnv, Receipt};
 
 #[tokio::main]
 async fn main() {
+    let _ = dotenvy::dotenv();
     // Initialize tracing. In order to view logs, run `RUST_LOG=info cargo run`
     tracing_subscriber::fmt()
         .with_env_filter(tracing_subscriber::filter::EnvFilter::from_default_env())
         .init();
 
-    // Start the server with the Risc0Prover
-    run_server::<Risc0Prover>().await;
+    let socket_addr = std::env::var("KAIROS_PROVER_SERVER_SOCKET_ADDR")
+        .expect("Failed to fetch environment variable KAIROS_PROVER_SERVER_SOCKET_ADDR");
+    let socket_addr = socket_addr
+        .parse::<std::net::SocketAddr>()
+        .expect("Failed to parse KAIROS_SERVER_SOCKET_ADDR");
+
+    let app = axum::Router::new()
+        .typed_post(prove_batch_route)
+        .with_state(());
+
+    tracing::info!("starting http server on `{}`", socket_addr);
+    let listener = tokio::net::TcpListener::bind(socket_addr).await.unwrap();
+    tracing::info!("listening on `{}`", socket_addr);
+    axum::serve(listener, app).await.unwrap()
 }
 
-pub struct Risc0Prover;
+#[derive(TypedPath, Debug, Clone, Copy)]
+#[typed_path("/api/v1/prove/batch")]
+pub struct ProveBatch;
 
-impl Prover for Risc0Prover {
-    type Error = String;
-    type ProofSystemData = Receipt;
+pub async fn prove_batch_route(
+    _: ProveBatch,
+    proof_inputs: Json<ProofInputs>,
+) -> Result<Json<(ProofOutputs, Receipt)>, (StatusCode, String)> {
+    let proof = tokio::task::spawn_blocking(move || prove_execution(proof_inputs.0))
+        .await
+        .map_err(|e| {
+            let e = format!("Error while joining proving task: {e}");
+            tracing::error!(e);
+            (StatusCode::INTERNAL_SERVER_ERROR, e)
+        })?;
 
-    fn prove_execution(
-        proof_inputs: kairos_circuit_logic::ProofInputs,
-    ) -> Result<Proof<Receipt>, String> {
-        let env = ExecutorEnv::builder()
-            .write(&proof_inputs)
-            .map_err(|e| format!("Error in ExecutorEnv builder write: {e}"))?
-            .build()
-            .map_err(|e| format!("Error in ExecutorEnv builder build: {e}"))?;
+    let proof = proof.map_err(|e| {
+        let e = format!("Error while proving batch: {e}");
+        tracing::error!(e);
+        (StatusCode::INTERNAL_SERVER_ERROR, e)
+    })?;
 
-        let receipt = risc0_zkvm::default_prover()
-            .prove(env, PROVE_BATCH_ELF)
-            .map_err(|e| format!("Error in risc0_zkvm prove: {e}"))?;
+    Ok(Json(proof))
+}
 
-        let logical_outputs = receipt
-            .journal
-            .decode()
-            .map_err(|e| format!("Error in receipt journal decode: {e}"))?;
+fn prove_execution(
+    proof_inputs: kairos_circuit_logic::ProofInputs,
+) -> Result<(ProofOutputs, Receipt), String> {
+    let env = ExecutorEnv::builder()
+        .write(&proof_inputs)
+        .map_err(|e| format!("Error in ExecutorEnv builder write: {e}"))?
+        .build()
+        .map_err(|e| format!("Error in ExecutorEnv builder build: {e}"))?;
 
-        receipt
-            .verify(PROVE_BATCH_ID)
-            .map_err(|e| format!("Error in risc0_zkvm verify: {e}"))?;
+    let receipt = risc0_zkvm::default_prover()
+        .prove(env, PROVE_BATCH_ELF)
+        .map_err(|e| format!("Error in risc0_zkvm prove: {e}"))?;
 
-        Ok(Proof {
-            logical_outputs,
-            proof_system_data: receipt,
-        })
-    }
+    let proof_outputs = receipt
+        .journal
+        .decode()
+        .map_err(|e| format!("Error in receipt journal decode: {e}"))?;
+
+    receipt
+        .verify(PROVE_BATCH_ID)
+        .map_err(|e| format!("Error in risc0_zkvm verify: {e}"))?;
+
+    Ok((proof_outputs, receipt))
 }
 
 #[cfg(test)]
@@ -61,9 +92,6 @@ mod tests {
             Withdraw,
         },
     };
-    use kairos_prover_server_lib::Prover;
-
-    use crate::Risc0Prover;
 
     #[test]
     fn test_prove_simple_batches() {
@@ -113,8 +141,8 @@ mod tests {
         ];
 
         test_prove_batch(batches, |proof_inputs| {
-            Ok(Risc0Prover::prove_execution(proof_inputs)
-                .map(|o| o.logical_outputs)
+            Ok(crate::prove_execution(proof_inputs)
+                .map(|(proof_outputs, _)| proof_outputs)
                 .expect("Failed to prove execution"))
         });
     }
